Hotfix: accessing cached values was refused if display parameter is passed
This fix is required because of a lack of conversion server ability to perform
successive transformations.
Add a comment in docstring to address explicitely this issue.

display and quality parameters are not handled explicitely by OOoDocument._convert
as they are just passed as they are to Image._convert .

Add a test to prove the fix