Rework 'mass workflow transition'
In the first dialog you still select which action to do on which objects, now
you can only select one action. The second dialog is the actual dialog from the
workflow rendered on the context of one document from documents grouped by
state and portal type. This means that if the workflow uses a custom dialog to
set workflow variables, this dialog will be shown and the variables will be
passed to all documents.
Of course some workflows dialog fields have no meaning in a multi-document
transition, that's why there is a mass_workflow_action=True request parameter
to detect this situation and behave accordingly in such case.