XMLMatrix: stop preserving order of coordinates for each axis
Renaming cells caused too much trouble:
- updateRelatedContent was not used
- there were potential catalog (fatal) errors like
  "uid of ... changed from ... (property) to ... (catalog, by path) !!!" or
  "uid of ... is ... and is already assigned to deleted in catalog !!!"

So this new implementation never changes ids of cells, and there may be holes
in allocated "partial ids".

getCellRange is also fixed to make sure we don't rely on the order of dict
items, as implemented by CPython. If it changed, matrices would be entirely
deleted.