Do not set namespace dict where exec is executed into as a default parameter dict.
When load() was called in checkConsistency where namespace_dict was not given
and thus equal to the default function parameter value, exec would set the
function namespace_dict parameter, thus the next call would have
namespace_dict already set when it was given.

This fixes testInvalidSourceCode when being ran for TestZodbExtensionComponent
and TestZodbDocumentComponent in the same runUnitTest.