Define versions and their priorities for ZODB Components on Business Template.
This is similar to portal skins priority, except that versions and priorities
are defined on ERP5Site (as 'VERSION | PRIORITY' where PRIORITY is a float) as
there is not one Folder for each version in contrary to portal skins.

On the Business Template itself, versions and priorities can be defined like
'VERSION | PRIORITY' or simply 'VERSION' (where PRIORITY is assumed to be
0.0). Upon installation, it will be merged into ERP5Site property in
descending order (by priority first and then by version for versions with the
same priority).