Fix convetsion and content type when convert to svg
In order to fix content type it was Introduced a small
Monkey-Patch for zope.contenttype in order to detect
image/svg+xml content types. This is broken even in newer
versions and this monkey patch is foward compatible.

Update Document and tests to permit conversions to SVG, this
prevents get Unauthorized when use ?format=svg.