Included some tests for SVG to PNG conversions
Using Image Document, this commit includes tests to assert
the conversion outcome when convert svg to png. The main goal
is assert if the image backgronud (external and embeded cases)
is incorpoted into the PNG generated.

The image comparation is very basic level, but enough for this tests.

It exposes an error when convert one svg that uses url with http.