Changes to fix test_store_as_event from testNotificationTool and testMailFromMailMessageEvent from testERP5Credential
after_path_and_method_id argument was added to not activate when Crededial
request object is not indexed yet. This is needed because when the method
searchAndActivate from catalog is called, if the object is not indexed, the
e-mail is not sent.

And extended script Entity_sendEmail to use from_url parameter when it is not
None. This case happens when the sender is a string instead of one object in
ERP5