Fix Trade Model Solver that wrongly skipped some TM related simulation movements
This fixes a regression in recent commit 6e758ade09d264daa8ec0b2da3888911f39843f0

The following line:

  simulation_movement_list = delivery_dict.get(movement) or \
    movement.getDeliveryRelatedValueList()

didn't work because the loop that fills delivery_dict does not necessarily
find all related simulation movements for every key.

New code is a little complicated because it tries to not call
getDeliveryRelatedValueList() when 'movement' is not related to trade model.