Prevent random failures when email table is not present
The main reason for the test_installBusinessTemplatesFromRepository_update_catalog
test is assert if the activities from created from clear catalog are
present on the activity lists or not.

This change re-creates the table using zsqls introduced by
erp5_ingestion_mysql_innodb_catalog to make sure the table is there when
the index activites related to the new BT5 be launched.