Do not recursively reindex security unnecessarily
In ERP5, objects that do not acquire local roles (according to their portal type configuration) do not need to be reindexed with their containers.
So do not reindex recursively if portal_types of contained objects don't need it.

Security reindexing can be customized with a type-based method.

Fix test that wrongly relied on unrestricted access for updating object roles.