Set max_rows 0 on IdTool_zGetValueList
So that we can get all the records of portal_ids table even when there are more
than 1000 id_group on the table. The ZSQLMethod was not intended to limit to
retrieve the records.
Of course it is better to get the records more gracefully like first getting
0..999, then 1000..1999. For now this is acceptable.
(Confirmed by Sebastien)