Add a method to update the persistent id_dict from portal_ids table.
This method is intended to use when we are disabled 'Store in the ZODB'
or set a big number as the 'Store interval', but at the same time we need to
update the persistent dict from the fresh portal_ids table at a certain point.

This method needs a unit test.