Remove magic columns special case.
If a column is explicitly resolved to a table, it cannot be resolved to
another later. This causes problems when writing a related key targeting
another table with the same column name.
This code duplicates work done in the (only) caller (in
SearchKey/RelatedKey.py), as it already registers the latest relation
table with the column to apply on it.