Avoid running out of file descriptors when backing up many objects in the trash upon bt5 installation.
Before, when installing a bt5 whose format has changed, many subobjects may
end up being backup up, but it keeps one FD per subobject, thus easily
reaching the limit of 1024 FDs (as given by ulimit -Sn/-Hn).

This is a followup of b71e5a7 specific to TrashTool as that commit seems to
only fix the issue when TrashTool is not used.