Extend grouping_reference with grouping_date
The grouping date is a property representing the date when this grouping is
effective.
In erp5_accounting, when setting a grouping reference, the grouping date will
be set to the latest date of grouped line. When running reports with
"omit_grouping_reference", lines that are grouped after the end date of the
report will still be displayed.

SQLCatalog structure changed, and can be updated with:
alter table catalog add column grouping_date datetime after grouping_reference;

You might also want to set a grouping date on movements with a grouping
reference in the migration.