Drop unneeded use of thread-local.
memcache module already handle thread-safety, and we actually want to be
transaction-local here, so use a volatile variable instead for caching
needs.
Removes the need to restart Zope for connection parameter changes to take
effect.
Fixes a bad interaction with SelectionTool's code, which causes
transactional isolation to be broken (just for this connector) and too many
network connections to be used.