Fix commit order of CMFActivity SQL connection on nodes with several zserver threads
When a ZODB connection is closed, it usually returns to a ZODB pool and may be
reused by another thread. If the SQL connection was open and is still in ZODB
cache, the _v_database_connection attribute is still there:
ActivityConnection.connect() is not called and a new instance of ZMySQLDA.db.DB
is created for the new thread without initializing its sort key.