ZMySQLDA: attach connections to ZODB connections instead of threads
This makes code simpler, faster and easier to understand.
It is easy to forget that ZODB connections can be reused by different threads,
which led to bug such as the one fixed by commit 2c11b76.

ZODB already maintains a pool of connections to reuse
so we don't need anymore to have one.