do not force section_category= except for draft
This does not work well with erp5_accounting_alternate_catalog.
It is left for draft transactions, because draft may not have any section set.

For the records, the idea behind is to prevent an unexpected behaviour when a
user have a section_category in his selection and calls a worklist which is
also supposed to display transactions which are not part of the selected
section_category. In such case, section_category= is passed to reset the
selected section_category and display all transactions, the problem is that
alternate catalog is not able to decide wether to display source or destination
section in that case