CatalogTool: support sorting on a related key
This approach consists in treating each keys from the default group (the
one for sort on and select dict) as a left join if the key supports it, at
column mapping level.

Also update testERP5Catalog, the specification changed a bit.
Before, using a related key in select_dict had the side effect of not selecting
documents without related documents unless the key is explicitly passed in
select_dict. Now all documents are always returned.