Fix inconsistency in Localizer and MessageCatalog translate method for message with no translation.
The former calls zope.i18n.translate() which calls MessageCatalog.translate()
by passing 'default' argument to the original message, thus returning the
original message when there is no translation. But, when the latter is called
without 'default' argument, it is set to message.strip(), thus returning the
strip'ed original message.