respect user's input for source or destionation in Base_addEvent.
before, user's input is completely ignored and calculated good values
are automatically set. and the commit fc54fbaad0 somehow tried to
respect user's source input, but it caused worse result for incoming
case. so this commit reverts fc54fbaad0 and set source and destination
just as user's input if not empty.