Refactor script to remove slow query that is not not so intuitive for end users.
To display the list of possible accounts we include validated nodes and  nodes that have been previously used on accounting movements. We include 2, because we want the user to be able to issue a report on a node that has been used in the past and is now invalidated.

The problem with this approach is that such query is very slow, when there are a lot of Movements.

Then, with the commit only validated accounts will be considered