erp5_promise: add script wich launch Fix Configuration Consitency
This script in order to be able to fix the site consistency via an url.
This script is present in erp5_promise because erp5_promise is one
of the bt5 installed by default. Fixing site consistency permits
to install others bt5 as erp5_configurator_standard.
This script is called externally via his url during the bootstrap
of scalability tests.