erp5_configurator_standard: add business_configuration_module object.
This new object has a bad id, we should change it. Also, may not be
in the good place in erp5_configurator_standard and should be
moved in another bt5.
The reason of his place in erp5_configurator_standard, is that
erp5_configurator_standard is automatically installed when
fix site consistency is done, thereby it install this preconfigured
object to. If this object is moved, the bt5 which will contains it
must be added to the list of bt5 to install when doing check site
consistency.