AccessTab: Fix an accesstab front page caching problem.
Fix a problem that access-tab-front-page wrongly caching old page, for example,
even if a user switching to use 'https' from 'http', old 'http' links are
remaining at the front page.

That was because the key of the CachingMethod which was used for caching the
page was ('user' and 'language').
Now the caching key is ('user', 'language' and 'site_url').