Add test to simulate the issue with the method TemplateTool.installBusinessTemplateListFromRepository
Using the method try to install Business Templates that are already
installed, the Business Templates is uploaded on ERP5 but not used. It
creates a lot of garbage because in the end you have many bt5s not
installed.

(cherry picked from commit 28e0f0a2217c8bae26707453eb269ef73039a2f3)