erp5_web: Take effective_date into account when looking content up.
effective_date, when defined, prevents a document from being visible even
if it is in the correct workflow state.
Note that this is orthogonal with workflow states: a document which must
not be visible should not be in a workflow state which makes it visible to
begin with. effective_date only affects website-style document listing and
lookup.