ZODB Components: Live Tests could only be ran once.
reload(), which should be called only for filesystem tests, was called for
ZODB Components. So on the first execution, sys.modules was updated with a new
instance of the module whereas the module in erp5.component.test stayed the
same. On the second execution, reload() failed it checks whether
sys.modules[MODULE] == erp5.component.test.MODULE.