runUnitTest: fixed weird "Duplicate tpc_begin calls for same transaction"
While launching an new portal component test, a "Duplicate tpc_begin
calls for same transaction" error was raised. This was coming from
the finally clause in runUnitTestList, in the commit of
ProcessingNodeTestCase.unregisterNode.

After long investigations, we found out that a new app was created
while the transaction was still opened. After more investigations,
it was discovered that the runUnitTestList code was raising an
exception, and this exception was hidden by the exception in the
finally close.

So this patch provides:
1 - a cancellation of the transaction to avoid "duplicate tpc_begin
    calls" error.
2 - a print of the exception before the finally clause is launched
    to make sure we have information on possible exceptions