CatalogTool: fix searchAndActivate when called with a condition on 'uid'
Because the method internally overrode 'uid' after the first group of objects,
passing a list of uids could make it go berserk and activate all objects in
catalog with an uid greater than those of the first group.

Although searchAndActivate was not initially written to be called with such
list, this commit fixes this instead of forbidding conditions on 'uid',
since it remains useful to easily split & group activities.