Remove username field server side autofill on login forms
<input login `request.get("__ac")` > removed

When the login form is rendered, the login field is not already filled.
We assume that the "__ac" cookie is deleted, so no need to generate the
login name server side. It also avoids to use old "__ac" cookies.
Browsers are already able to autocomplete them just by clicking on the field.