erp5_test_result: use activeSense method of alarm instead of calling script
The activeSense method is doing more than calling the alarm script, it also change
the user. Without using activeSense, the alarm script will be called under the user
doing something on test node/ test suite, and this could conduct to different results
than if it is run with the nobody user.