TemplateTool.upgradeSite: allow to define automatic catalog updating while keeping standard behaviour.
fixes promises (full_text_table not available). Default behaviour of
upgradeSite is not changed. Updating catalog should be handled outside
upgradeSite. The update_catalog argument is only introduced in this
commit so that it can be propagated to updateBusinessTemplateFromUrl in
case one does not know which business templates need catalog updating.