Add ERP5Security Dumb HTTP Extraction Plugin
This the default way to authenticate in ERP5. Without this,
it is DANGEROUS to activate another plugin.

The PluggableAuthenticationService lists all activated plugin
and if at least one is available then it deactivate the default
way to authenticate. And if you are not able to pass the activated
plugin then you cannot login anymore.

To prevent this, we can activate a plugin which uses the default
way to authenticate (Dumb HTTP Extraction Plugin), and THEN activate
another plugin.