CMFActivity: in case of ConflictError, retry earlier than for other failures
This tweaks retry delays as follows (seconds):

        ConflictError  other failures (K = 1 + retry², with retry >= 0)
before     30             15 * K
after      15             30 * K

Today, bigger delays in case of errors should not be an issue because the
quality of ERP5 has improved a lot and normal code should not rely anymore on
this.

We also don't want to lower ConflictError delay too much, because this
increase the probability of conflicts.

This will be required to improve invokeGroup, in case that a message fails.
We'd like that:
- successful messages are retried immediately
- skipped messages are retried next, and separately
- at last, failed messages are retried, also separately